Move api specific test to test_api
