+++ emmaa_service/test_api.py
@@ -1,11 +1,41 @@
 import unittest
 
-from .api import app
+from moto import mock_s3
+from emmaa.util import make_date_str
+from emmaa.tests.test_s3 import setup_bucket, TEST_BUCKET_NAME
+
+
+@mock_s3
+def test_api_load_from_s3():
+    from emmaa_service.api import is_available, get_latest_available_date, \
+        _get_test_corpora, _get_model_meta_data, get_model_config
+    from emmaa.model import last_updated_date
+    client = setup_bucket(add_model=True, add_model_stats=True,
+                          add_test_stats=True)
+    today = make_date_str()[:10]
+    other_day = '2020-01-01'
+    assert is_available('test', 'simple_tests', today, TEST_BUCKET_NAME)
+    assert not is_available(
+        'test', 'large_corpus_tests', today, TEST_BUCKET_NAME)
+    assert not is_available(
+        'test', 'simple_tests', other_day, TEST_BUCKET_NAME)
+    assert get_latest_available_date(
+        'test', 'simple_tests', bucket=TEST_BUCKET_NAME) == today
+    config = get_model_config('test', TEST_BUCKET_NAME)
+    assert config
+    test_corpora = _get_test_corpora('test', TEST_BUCKET_NAME)
+    assert test_corpora == {'simple_tests'}
+    metadata = _get_model_meta_data(bucket=TEST_BUCKET_NAME)
+    assert len(metadata) == 1
+    assert len(metadata[0]) == 2
+    assert metadata[0][0] == 'test'
+    assert metadata[0][1] == config
 
 
 class EmmaaApiTest(unittest.TestCase):
 
     def setUp(self):
+        from .api import app
         app.testing = True
         self.app = app.test_client()
 
@@ -23,7 +53,7 @@ class EmmaaApiTest(unittest.TestCase):
                       'query': {'objectSelection': 'ERK',
                                 'subjectSelection': 'BRAF',
                                 'typeSelection': 'activation'},
-                      'register': 'false',
+                      'register': False,
                       'test': 'true'}
 
         resp = self.app.post('/query/submit', json=test_query)
@@ -40,7 +70,7 @@ class EmmaaApiTest(unittest.TestCase):
                           'query': {'objectSelection': 'AKT1',
                                     'subjectSelection': 'PIK3CA',
                                     'typeSelection': 'activation'},
-                          'register': 'false'}
+                          'register': False}
 
         resp = self.app.post('/query/submit', json=test_query)
